[Silver IV] Title: 다음 소수, Time: 360 ms, Memory: 11952 KB

+++ 백준/Silver/4134. 다음 소수/다음 소수.java
@@ -5,7 +5,7 @@ public class Main {
 	private static boolean isPrime(long n) {
 		if(n <= 1) return false;
 		
-		for(long i = 2; i <= Math.sqrt(n); i++) {
+		for(int i = 2; i <= Math.sqrt(n); i++) {
 			if(n % i == 0) return false;
 		}
